Renamed ChangeQueue to LinearChangeQueue: prepare to create base change queue interface
