+++ undofx/src/main/java/org/fxmisc/undo/DirectedAcyclicUndoGraph.java
@@ -1,11 +1,8 @@
 package org.fxmisc.undo;
 
-import org.fxmisc.undo.impl.ChangeQueue;
 import org.fxmisc.undo.impl.NonLinearChange;
 import org.fxmisc.undo.impl.NonLinearChangeQueue;
 import org.reactfx.SuspendableNo;
-
-import java.util.function.Supplier;
 
 public interface DirectedAcyclicUndoGraph<S extends NonLinearChangeQueue<C>, C> {
