+++ undofx/src/test/java/org/fxmisc/undo/impl/ZeroSizeLinearChangeQueueTest.java
@@ -2,14 +2,14 @@ package org.fxmisc.undo.impl;
 
 import static org.junit.Assert.*;
 
-import org.fxmisc.undo.impl.ChangeQueue.QueuePosition;
+import org.fxmisc.undo.impl.LinearChangeQueue.QueuePosition;
 import org.junit.Test;
 
 public class ZeroSizeLinearChangeQueueTest {
 
     @Test
     public void testPositionValidityOnOverflow() {
-        ChangeQueue<Integer> queue = new ZeroSizeLinearChangeQueue<>();
+        LinearChangeQueue<Integer> queue = new ZeroSizeLinearChangeQueue<>();
         QueuePosition pos0 = queue.getCurrentPosition();
         assertTrue(pos0.isValid());
         queue.push(1);
